Do not run setup-wizard if it ran earlier on
